requirements.txt
python-telegram-bot==21.3
httpx==0.27.0
apscheduler==3.10.4
